Remove SymmetrizedWordAlignmentEngine
